Merge nickbit to backend

+++ backend/mall-common/src/main/java/org/frogsoft/mall/common/model/comment/Comment.java
@@ -34,7 +34,7 @@ public class Comment {
 
     private int type;
 
-    private int parentId;
+    private Long parentId;
 
     private LocalDateTime commentTime;
